+++ src/time_unit/days_of_week.rs
@@ -1,0 +1,37 @@
+use ::schedule::{Ordinal, OrdinalSet, ExpressionError};
+use ::time_unit::TimeUnitField;
+use std::borrow::Cow;
+
+pub struct DaysOfWeek(OrdinalSet);
+
+impl TimeUnitField for DaysOfWeek {
+  fn from_ordinal_set(ordinal_set: OrdinalSet) -> Self {
+    DaysOfWeek(ordinal_set)
+  }
+  fn name() -> Cow<'static, str> {
+    Cow::from("Days of Week")
+  }
+  fn inclusive_min() -> Ordinal {
+    1
+  }
+  fn inclusive_max() -> Ordinal {
+    7
+  }
+  fn ordinal_from_name(name: &str) -> Result<Ordinal, ExpressionError> {
+    //TODO: Use phf crate
+    let ordinal = match name.to_lowercase().as_ref() {
+      "sun" | "sunday" => 1,
+      "mon" | "monday" => 2,
+      "tue" | "tues" | "tuesday" => 3,
+      "wed" | "wednesday" => 4,
+      "thu" | "thurs" | "thursday" => 5,
+      "fri" | "friday" => 6,
+      "sat" | "saturday" => 7,
+      _ => return Err(ExpressionError(format!("'{}' is not a valid day of the week.", name)))
+    };
+    Ok(ordinal)
+  }
+  fn ordinals(&self) -> &OrdinalSet {
+    &self.0
+  }
+}
