+++ tests/lib.rs
@@ -3,29 +3,30 @@ extern crate chrono;
 
 #[cfg(test)]
 mod tests {
-  use cron::CronSchedule;
+  use cron::Schedule;
+  use std::str::FromStr;
   use chrono::*;
 
   #[test]
-  fn test_parse_with_seconds() {
+  fn test_parse_with_year() {
     let expression = "1 2 3 4 5 6 2015";
-    assert!(CronSchedule::parse(expression).is_ok());
+    assert!(Schedule::from_str(expression).is_ok());
   }
 
   #[test]
   fn test_parse_with_seconds_list() {
-    let expression = "1,30,40 2 3 4 5 6 2015";
-    assert!(CronSchedule::parse(expression).is_ok());
+    let expression = "1,30,40 2 3 4 5 Mon-Fri";
+    assert!(Schedule::from_str(expression).is_ok());
   }
   
   #[test]
   fn test_parse_with_lists() {
-    let expression = "1 2,17,51 1-3,6,9-11 4,29 2,3,7 6 2015-2017";
-    let schedule = CronSchedule::parse(expression).unwrap();
+    let expression = "1 2,17,51 1-3,6,9-11 4,29 2,3,7 Tues";
+    let schedule = Schedule::from_str(expression).unwrap();
     let mut date = UTC::now();
     println!("Fire times for {}:", expression);
     for _ in 0..20 {
-      date = schedule.next_utc_after(&date).expect("No further dates!");
+      date = schedule.after(&date).next().expect("No further dates!");
       println!("-> {}", date);
     }
     assert!(true);
@@ -33,38 +34,38 @@ mod tests {
 
   #[test]
   fn test_upcoming_iterator() {
-    let expression = "2,17,51 1-3,6,9-11 4,29 2,3,7 6 2015-2017";
-    let schedule = CronSchedule::parse(expression).unwrap();
+    let expression = "0 2,17,51 1-3,6,9-11 4,29 2,3,7 Wed";
+    let schedule = Schedule::from_str(expression).unwrap();
     println!("Upcoming fire times for '{}':", expression);
-    for datetime in schedule.upcoming().take(12) {
+    for datetime in schedule.upcoming(UTC).take(12) {
       println!("-> {}", datetime);
     }
     assert!(true);
   }
 
   #[test]
-  fn test_parse_without_seconds() {
-    let expression = "1 2 3 4 5 2015";
-    assert!(CronSchedule::parse(expression).is_ok());
+  fn test_parse_without_year() {
+    let expression = "1 2 3 4 5 6";
+    assert!(Schedule::from_str(expression).is_ok());
   }
   
   #[test]
   fn test_parse_too_many_fields() {
-    let expression = "1 2 3 4 5 6 7 8 9 2015";
-    assert!(CronSchedule::parse(expression).is_err());
+    let expression = "1 2 3 4 5 6 7 8 9 2019";
+    assert!(Schedule::from_str(expression).is_err());
   }
   
   #[test]
   fn test_not_enough_fields() {
-    let expression = "1 2 3 2015";
-    assert!(CronSchedule::parse(expression).is_err());
+    let expression = "1 2 3 2019";
+    assert!(Schedule::from_str(expression).is_err());
   }
 
   #[test]
   fn test_next_utc() {
-    let expression = "1 2 3 4 10 6 2015,2016,2017,2018";
-    let schedule = CronSchedule::parse(expression).unwrap();
-    let next = schedule.next_utc().expect("There was no upcoming fire time.");
+    let expression = "1 2 3 4 10 Fri";
+    let schedule = Schedule::from_str(expression).unwrap();
+    let next = schedule.upcoming(UTC).next().expect("There was no upcoming fire time.");
     println!("Next fire time: {}", next.to_rfc3339());
   }
 }
